feat: improve node accessors

+++ examples/maze/world.cpp
@@ -2,5 +2,5 @@
 
 World::World(Engine* pEngine, int size=11): GameObject(pEngine), sideSize(size) {
   data.clear();
-  data.resize((size+1)*(size+1)*2);
+  data.resize((size_t)(size+1)*(size+1)*2);
 }
